[:recycle:]: Update theme and add worker
