+++ src/components/AsciinemaPlayer/index.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useRef } from 'react';
 import * as AsciinemaPlayerLibrary from 'asciinema-player';
 import 'asciinema-player/dist/bundle/asciinema-player.css';
 
-type AsciinemaPlayerProps = {
+export interface AsciinemaPlayerProps {
   src: string;
   cols: string;
   rows: string;
@@ -18,15 +18,12 @@ type AsciinemaPlayerProps = {
   fontSize: string;
 };
 
-function AsciinemaPlayer({
-  src,
-  ...asciinemaOptions
-}: React.FC<AsciinemaPlayerProps>): JSX.Element {
+function AsciinemaPlayer({src, ...asciinemaOptions}: AsciinemaPlayerProps): JSX.Element {
   const ref = useRef<HTMLDivElement>(null);
   useEffect(() => {
     const currentRef = ref.current;
     AsciinemaPlayerLibrary.create(src, currentRef, asciinemaOptions);
-  }, [src, asciinemaOptions]);
+  }, [src]);
 
   return <div ref={ref} />;
 }
