+++ src/pages/index.module.css
@@ -6,32 +6,9 @@
   overflow: hidden;
 }
 
-.videocontainer {
-  padding: 4rem 0;
-  text-align: justify;
-  position: relative;
-  overflow: hidden;
-}
-
-.features {
-  display: block;
-  align-items: center;
-  width: 100%;
-}
-
-.asciicasts {
-  display: block;
-  align-items: center;
-  width: 100%;
-}
-
 @media only screen and (max-width: 996px) {
   .herobanner {
     padding: 1rem 0;
-  }
-
-  .videocontainer {
-    padding: 2rem 0;
   }
 }
 
@@ -109,18 +86,18 @@ div[class^="announcementBarContent"] a:hover {
 
 .hero {
   background-color: #2b3137;
-  padding: 48px;
+  padding: 4rem;
 }
 
 .heroInner {
   margin: 0 auto;
   max-width: 1100px;
-  padding: 0 20px;
+  padding: 0 2rem;
 }
 
 .heroProjectTagline {
   color: #fff;
-  font-size: 60px;
+  font-size: 3.6rem;
   margin: 0;
 }
 
@@ -145,7 +122,7 @@ div[class^="announcementBarContent"] a:hover {
 
   to {
     opacity: 1;
-    transform: scale(1);
+    transform: scale(0.8, 0.8);
   }
 }
 
@@ -156,10 +133,11 @@ div[class^="announcementBarContent"] a:hover {
   margin-top: 20px;
   padding: 0 20px 20px;
   height: auto;
+  width: 200px;
 }
 
 .indexCtas {
-  --ifm-button-size-multiplier: 1.4;
+  --ifm-button-size-multiplier: 1.2;
 
   display: flex;
   flex-wrap: wrap;
@@ -181,8 +159,8 @@ div[class^="announcementBarContent"] a:hover {
 
 @media only screen and (max-width: 768px) {
   .hero {
-    padding-left: 20px;
-    padding-right: 20px;
+    padding-left: 1.8rem;
+    padding-right: 1.8rem;
   }
 
   .heroInner {
@@ -190,7 +168,7 @@ div[class^="announcementBarContent"] a:hover {
   }
 
   .heroProjectTagline {
-    font-size: 36px;
+    font-size: 2.2rem;
     text-align: center;
   }
 
@@ -198,10 +176,13 @@ div[class^="announcementBarContent"] a:hover {
     display: block;
     float: none;
     margin: 0 auto;
+    width: 140px;
   }
 
   .indexCtas {
     justify-content: center;
+
+    --ifm-button-size-multiplier: 1;
   }
 
   .indexCtas a {
